Fix flow layout warnings
Some warnings were encountered on homepages when scrolling, and due to the fact that a large cell was reused for a smaller layout. In the end the layout was correct, but since a larger layout would briefly be forced to a smaller one, the layout would break. To fix this, it suffices to clear the collection on reuse (which was not the case)