Do not use invalidation contexts
Idea is good, but the layout is already invalidated when scrolling, so this brings additional complexity without real benefits (invalidation every 10 seconds).

If we run into performance issues we could have better invalidation strategies for channel headers and the vertical indicator, but in this case I would recommend we have a single prepare implementation which checks the invalidation context and only recalculates layout constants, rather than some mutated state.